Moved UtilFineFileTest to logic.util.io
